Task_5: added scalable application structure, new module, some semantic tags, flex basis/grow/shrink, alias and small refactoring
