+++ my-app/src/app/app.component.scss
@@ -1,63 +1,16 @@
+@import '../assets/variables';
+
 :host {
   display: flex;
   flex-direction: column;
   min-height: 100vh;
 
-  app-section, app-courses, app-load-more-button, .no-data-container {
-    display: block;
-    width: 80%;
-    margin: auto;
+  app-header, app-breadcrumbs, app-footer {
+    flex: 0 0 $headerHeight;
   }
 
-  form, .input-container {
-    display: inline-flex;
-  }
-
-  form {
-    margin: 1% 0 0 10%;
-
-    .input-container {
-      background-color: white;
-      border-radius: 5px;
-
-      .search-courses-input {
-        border: none;
-        outline: none;
-      }
-
-      .search-courses-input, .search-icon {
-        align-self: center;
-        margin-left: 10px;
-      }
-    }
-
-    .search-button {
-      padding: 7px;
-      margin-left: 5px;
-      background-color: limegreen;
-    }
-  }
-
-  .add-course-button-container, .add-course-button {
-    display: flex;
-  }
-
-  .add-course-button-container {
-    justify-content: flex-end;
-    margin-right: 10%;
-
-    .add-course-button {
-      padding: 5px;
-      background: cornflowerblue;
-    }
-  }
-
-  .no-data-container {
-    margin-bottom: 70vh;
-    border-radius: 20px;
-    background-color: white;
-    line-height: 40px;
-    text-align: center;
+  app-courses-page {
+    flex: 1 1 calc(100vh - (3*$headerHeight));
   }
 }
