+++ my-app/src/app/components/footer/footer.component.html
@@ -1,1 +1,1 @@
-<p>Lorem ipsum dolor sit amet.</p>
+<footer>Lorem ipsum dolor sit amet.</footer>
