+++ my-app/src/app/components/footer/footer.component.scss
@@ -1,9 +1,11 @@
+@import '../../../assets/variables';
+
 :host {
   display: flex;
   justify-content: flex-end;
   padding-right: 10%;
   color: lightgray;
   font-size: 10px;
-  line-height: 40px;
+  line-height: $headerHeight;
   background-color: black;
 }
