+++ my-app/src/app/components/header/header.component.html
@@ -1,4 +1,4 @@
-<div class="header-container">
+<header class="header-container">
   <div class="logo-container">
     <i class="fa-solid fa-square-caret-right fa-2x"></i>
     <p class="logo-text">VIDEO COURSE</p>
@@ -13,4 +13,4 @@
       <p class="authorization-text">Log off</p>
     </button>
   </div>
-</div>
+</header>
