+++ my-app/src/app/components/header/header.component.scss
@@ -10,7 +10,7 @@
 .header-container {
   align-items: center;
   justify-content: space-between;
-  height: 40px;
+  height: 100%;
   background-color: black;
 
   .logo-container {
