+++ my-app/src/app/components/index.ts
@@ -1,0 +1,3 @@
+export * from './header/header.component';
+export * from './breadcrumbs/breadcrumbs.component';
+export * from './footer/footer.component';
