+++ my-app/src/app/modules/courses-page/components/courses/courses.component.scss
