+++ my-app/src/app/modules/courses-page/components/courses/courses.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input, Output, EventEmitter, ChangeDetectionStrategy} from '@angular/core';
 
-import {Course} from "../../shared";
+import {Course} from "@shared/";
 
 @Component({
   selector: 'app-courses',
