+++ my-app/src/app/modules/courses-page/components/index.ts
@@ -1,0 +1,2 @@
+export * from './courses/courses.component';
+export * from './load-more-button/load-more-button.component';
