+++ my-app/src/app/modules/courses-page/core/directives/is-course-fresh.directive.ts
