+++ my-app/src/app/modules/courses-page/core/index.ts
@@ -1,0 +1,4 @@
+export * from './pipes/duration-handler/duration-handler.pipe';
+export * from './pipes/filter/filter.pipe';
+export * from './pipes/orderBy/order-by.pipe';
+export * from './directives/is-course-fresh.directive';
