+++ my-app/src/app/modules/courses-page/core/pipes/duration-handler/duration-handler.pipe.spec.ts
