+++ my-app/src/app/modules/courses-page/core/pipes/orderBy/order-by.pipe.ts
@@ -2,7 +2,7 @@ import {Pipe, PipeTransform} from '@angular/core';
 
 import orderBy from "lodash/orderBy";
 
-import {Course} from "../../types";
+import {Course} from "@shared/";
 
 @Pipe({
   name: 'orderBy'
