+++ my-app/src/app/modules/courses-page/courses-page.component.scss
@@ -1,0 +1,66 @@
+@import '../../../assets/variables';
+
+:host {
+  display: block;
+  width: 80%;
+  margin: auto;
+
+  app-courses, app-load-more-button, .no-data-container {
+    display: block;
+    margin: auto;
+  }
+
+  .form-container {
+    margin-top: 15px;
+
+
+    form, .input-container {
+      display: inline-flex;
+    }
+
+    form {
+      .input-container {
+        background-color: white;
+        border-radius: 5px;
+
+        .search-courses-input {
+          border: none;
+          outline: none;
+        }
+
+        .search-courses-input, .search-icon {
+          align-self: center;
+          margin-left: 10px;
+        }
+      }
+
+      .search-button {
+        padding: 7px;
+        margin-left: 5px;
+        background-color: limegreen;
+      }
+    }
+  }
+
+  .add-course-button-container, .add-course-button {
+    display: flex;
+  }
+
+  .add-course-button-container {
+    justify-content: flex-end;
+    margin-bottom: 15px;
+
+    .add-course-button {
+      padding: 5px;
+      background: cornflowerblue;
+    }
+  }
+
+  .no-data-container {
+    margin-bottom: 70vh;
+    border-radius: 20px;
+    background-color: white;
+    line-height: $headerHeight;
+    text-align: center;
+  }
+}
