+++ my-app/src/app/modules/index.ts
@@ -1,0 +1,1 @@
+export * from './courses-page/courses-page.module';
