+++ my-app/src/app/shared/index.ts
@@ -1,5 +1,2 @@
-export * from './pipes/duration-handler/duration-handler.pipe';
-export * from './pipes/filter/filter.pipe';
-export * from './pipes/orderBy/order-by.pipe';
-export * from './directives/is-course-fresh.directive';
-export * from './types';
+export * from './models/course.model';
+export * from './models/user.model';
