+++ my-app/src/app/shared/models/user.model.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: number;
+  firstName: string;
+  lastName: string;
+}
