+++ my-app/src/assets/variables.scss
@@ -1,0 +1,1 @@
+$headerHeight: 40px;
